Fix setting SIL_DEBUG_ENABLE to 0
When SIL_DEBUG_ENABLE is set to 0 gcc complains a lot about 'set but
unused variables'. To make sure the compiler can see the variables are
used, move the use of SIL_DEBUG_ENABLE in the C code rather than CPP. The
resulting code is the same, as the compiler optimizes code that it will
never reach.

Closes #1311

Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>