Ccx.c: Use global to sync threads
Passing it via a struct is a bit awkward as the struct needs to be in
sync with assembly definitions. Using a global makes it easier as the
assenbly code has a direct reference to it.

Also RestoreResetVector() check for having launched all APs is dropped
as getting out of the loop for launching Ap already means that this
happened.

Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>